// apps/web/src/lib/api/generated/types/updateUserDtoRoleName.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Linktal Recruitment API
 * API for Linktal Recruitment OS
 * OpenAPI spec version: 1.0
 */

export type UpdateUserDtoRoleName = typeof UpdateUserDtoRoleName[keyof typeof UpdateUserDtoRoleName];


export const UpdateUserDtoRoleName = {
  admin: 'admin',
  manager: 'manager',
  consultant: 'consultant',
  finance: 'finance',
  researcher: 'researcher',
  viewer: 'viewer',
} as const;